+++ Models/DatabaseDiagnosticsInfo.cs
@@ -31,8 +31,9 @@ public class DatabaseDiagnosticsInfo
     public int ReputationCount { get; set; }
     public int WarehouseItemCount { get; set; }
     public int WikiItemCount { get; set; }
+    public int PilotProfileCount { get; set; }
 
-    public int TotalRecordsCount => SessionCount + EventCount + ContractCount + FleetShipCount + PoiCount + ReputationCount + WarehouseItemCount + WikiItemCount;
+    public int TotalRecordsCount => SessionCount + EventCount + ContractCount + FleetShipCount + PoiCount + ReputationCount + WarehouseItemCount + WikiItemCount + PilotProfileCount;
 
     // Tabellen- & Struktur-Validierung
     public List<string> ExistingTables { get; set; } = new();